Updated reader for unit tests
